+++ lib/src/extensions/byte_stream.dart
@@ -14,6 +14,7 @@
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:async/async.dart';
 import 'package:charcode/charcode.dart';
@@ -59,15 +60,24 @@ extension ByteStreamExtensions on Stream<List<int>> {
   /// This treats [this] as a process that emits only stdout. If [this] emits an
   /// error, it's treated the same as an unhandled Dart error in a
   /// [Script.capture] block: it's printed to stderr and the virtual stream
-  /// process exits with error code 257.
+  /// process exits with error code `257`.
+  ///
+  /// [Script.kill] returns `true` if the stream was interrupted and the script
+  /// exits with [Script.exitCode] `143`, or `false` if the stream was already
+  /// closed.
   Script get _asScript {
+    var signalCloser = StreamCloser<List<int>>();
     return Script.fromComponents("stream", () {
       var exitCodeCompleter = Completer<int>.sync();
       return ScriptComponents(
           NullStreamSink(),
-          onDone(() => exitCodeCompleter.complete(0)),
+          transform(signalCloser).onDone(() =>
+              exitCodeCompleter.complete(signalCloser.isClosed ? 143 : 0)),
           Stream.empty(),
           exitCodeCompleter.future);
+    }, onSignal: (_) {
+      signalCloser.close();
+      return true;
     });
   }